requirements.txt
altgraph==0.17.4
anyio==4.3.0
APScheduler==3.6.3
backports-datetime-fromisoformat==2.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
common==0.1.2
cryptography==42.0.5
decorator==4.4.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
m3u8~=4.1.0
moviepy==1.0.3
Naked==0.1.32
numpy==1.26.4
packaging==24.1
pefile==2023.2.7
pillow==10.3.0
proglog==0.1.10
pycparser==2.22
pycryptodome==3.20.0
python-dotenv~=1.0.1
python-telegram-bot
pyrogram
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests~=2.31.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.1
tornado==6.4
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
vk-api==11.9.9
